moved from roomfinder to navigatum

+++ app/src/main/java/de/tum/in/tumcampusapp/api/navigatum/model/details/NavigationCordsDto.kt
@@ -1,6 +1,6 @@
 package de.tum.`in`.tumcampusapp.api.navigatum.model.details
 
 data class NavigationCordsDto(
-    var lat: Float = 0f,
-    var lon: Float = 0f,
+    var lat: Double = 0.0,
+    var lon: Double = 0.0
 )
